ATL call crash fix
AgpsStatus added with a couple string fields for wifi
which are not NULL ended when they are not used. This
get passed to AFW JNI, and allow UTF8 string with the
open ended buffer with junks in it. It might cause
exception depending on the content of the garbadge.

CRs-Fixed: 405473

Change-Id: If762748a4c63fe1be59491b5d08dbadd479176b3