gps: Add A-GLONASS Position protocol in gps.conf
OEMs rely on gps.conf to config gps features. We want to have
default config in the conf file however to both serve as
recommendeded setting and as setting template. This added
A-GLONASS protocol is already supported by the current
implementation. Now we are adding the default config to the
conf file."

Change-Id: Ib9f66d01963c87ac5b3d70a68e1d4dd085771d47
CRs-fixed: 454855