IPACM: speed up to construct wifi-client RT rules
For client connect/disconnect case, if not in the
bridge mode, the new_neigh for wifi-client only
comes when client gets assigned ipv4 and delayed
the IPACM to construct routing rules for it. The
fix is to first construct RT rules based on previous
ipv4 when client connected to AP, and deleted,
re-added if ipv4 assignment is changed.

Change-Id: I165a5eb1b981a8db162b8205350dd88e97069651