ipa: check return status of HAL calls
If the underlying HAL service object is invalidated, it
might result in a HAL transaction failure and the ipacm
process would be forced to abort.

To prevent the transaction error, use the isOk check right
after the HAL API is invoked to check whether the HAL
transaction was successful or not.

Change-Id: I4d71b35b6a31a6b705c1f61b542e1e82d1716a21
CRs-Fixed: 2340020